feat(agreements): only set filter other than all
